Support: provide anchors
I considered text fragments too, e.g.

    https://rust-for-linux.com/industry-and-academia-support#:~:text=“Samsung,stack.”

which work fine in most browsers, but not Firefox.

I considered `div`s around the quote, but I did not want to depend on
the renderer to do the right thing, even if it currently worked. Moreover,
it makes the hightlighting feature in the next commit a bit more tricky
and complicates the DOM tree. Finally, we may want to e.g. add logos anyway
later on for the quotes, so just do HTML for this.

Suggested-by: Andreas Hindborg <a.hindborg@samsung.com>
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>