Contributing: clarify what the exceptions are intended to cover
In a recent discussion around a networking submission, it was
pointed out that "exceptions" could be more specific [1]. In other
words, the "to simplify the upstreaming process" wording in
the webpage may not be sufficient.

Thus clarify it further with similar notes/examples as those given
in the Zulip discussion.

In addition, link to the recently added "Out-of-tree modules" page
for further details on those.

Link: https://rust-for-linux.zulipchat.com/#narrow/stream/396592-Networking/topic/NAPI.20Abstractions/near/386129495 [1]
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>