Add "Out-of-tree modules" page
In a recent discussion around a networking submission, it seemed
that some extra clarifications about out-of-tree modules could
be useful for developers looking to develop Rust code for
the kernel [1].

Thus add an "Out-of-tree modules" page to clarify the focus of
the Rust for Linux project, what is intended to be supported
around out-of-tree code and what kinds of patches make sense to
submit to the kernel mailing list.

Importantly, add links to other pages from the mainline kernel
documentation.

Link: https://rust-for-linux.zulipchat.com/#narrow/stream/396592-Networking/topic/NAPI.20Abstractions/near/385885376 [1]
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>