Improve "AMCC QT2025 PHY Driver" description
Refine the description of the AMCC QT2025 driver to improve clarity
and accuracy. The original wording suggested that the PHY abstraction
was introduced along with the driver, while the intended meaning was
that the driver was merged together with improvements to the existing
PHY abstraction.

Note that the PHY abstraction was merged with the ASIX PHY driver.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>