Update Tyr Upstream Development Process
Currently the upstream development process for Tyr suggests that a
developer could open a merge request for review and discussion in
the panfrost-linux repository on Freedesktop GitLab.

While merge requests can still be a useful collaboration tool, other DRM
developers might not be reading those merge requests, so to be absolutely
clear, just direct new upstream contributions to the mailing lists.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>